backend/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1

# Data Processing
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2

# Visualization
plotly==5.19.0
kaleido==0.2.1

# PDF Generation
weasyprint==62.3
pydyf==0.10.0
jinja2==3.1.3

# AI Integration
openai==1.13.3

# Email
python-multipart==0.0.9
resend==2.0.0

# Environment & Configuration
python-dotenv==1.0.1

# HTTP Client
requests==2.31.0
httpx==0.27.0

# Async Support
asyncio-mqtt==0.16.2

# Caching (Redis support)
redis==5.0.3

# Database (for future expansion)
sqlalchemy==2.0.28
alembic==1.13.1

# Security
cryptography==42.0.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0

# Development
black==24.2.0
isort==5.13.2
mypy==1.8.0

# Monitoring & Logging
structlog==24.1.0
sentry-sdk[fastapi]==1.40.6

# Rate Limiting
slowapi==0.1.9 